+++ src/canister/src/migration/read_stores.rs
@@ -6,7 +6,7 @@ use super::old_models::profile_models::Profile as OldProfile;
 
 use super::call_wrappers::{
     read_attendees_entries, read_events_entries, read_groups_entries, read_members_entries,
-    read_notifications_entries, read_profile_entries,
+    read_profile_entries,
 };
 
 use candid::{CandidType, Deserialize};
@@ -18,7 +18,7 @@ pub struct OldData {
     pub old_profiles: Vec<(String, OldProfile)>,
     pub old_events: Vec<(String, OldEvent)>,
     pub old_event_attendees: Vec<(String, OldAttendee)>,
-    pub old_notifications: Vec<(u32, String)>,
+    // pub old_notifications: Vec<(u32, String)>,
 }
 
 pub async fn read_old_data(env: String) -> OldData {
@@ -27,7 +27,7 @@ pub async fn read_old_data(env: String) -> OldData {
     let old_profiles = read_profile_entries(env.clone()).await;
     let old_events = read_events_entries(env.clone()).await;
     let old_event_attendees = read_attendees_entries(env.clone()).await;
-    let old_notifications = read_notifications_entries(env.clone()).await;
+    // let old_notifications = read_notifications_entries(env.clone()).await;
 
     OldData {
         old_members,
@@ -35,6 +35,6 @@ pub async fn read_old_data(env: String) -> OldData {
         old_profiles,
         old_events,
         old_event_attendees,
-        old_notifications,
+        // old_notifications,
     }
 }